catelog payload created in sp-api amazon

+++ src/main/java/io/github/dft/amazon/model/catalogitems/v202204/ItemImage.java
@@ -9,6 +9,6 @@ public class ItemImage {
 
     private String variant;
     private String link;
-    private Integer height;
-    private Integer width;
+    private Double height;
+    private Double width;
 }
